+++ CHANGELOG.md
@@ -7,14 +7,15 @@ All notable changes to this project will be documented here.
 - Have a algorithm to clean gpx files
 - Improve predictions with data analyses
 - Improve slope dependencies with speed
-- Temperature can be set on every pages
 - Set a french/english flag
 - Add language menu
 
-#### [2.4.1] - 2022-09-23
+#### [2.4.3] - 2022-09-23
 
+- Temperature can be set on every pages
 - Fix weight input
 - Add axis on graphs
+- Start introducing fr/en translations
 
 #### [2.3.2] - 2022-09-22